Renamed classes: Card->CardGroup, calculator->Calculator.
Changed the way the groups are stored in Calculator.h and the methods to add and remove groups.

Added some functions in main.cpp to check if the code is working, the probability is still slightly off, it needs to be checked again.

ToBeFixed: calculate methods in Calculator.cpp and need to add some ifs to check the outofbounds exceptions.